material-ui + cleanup
