--- src/App.tsx
@@ -1,14 +1,0 @@
-import React, { Component, ReactNode } from 'react';
-import Board from './board';
-
-class App extends React.Component<{}, {}> {
-  render(): ReactNode {
-    return (
-        <div>
-          <Board />
-        </div>
-    );
-  }
-}
-
-export default App;
