+++ src/index.tsx
@@ -1,7 +1,18 @@
-import React from 'react';
+import React, { Component, ReactNode } from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
 import * as serviceWorker from './serviceWorker';
+import Board from './board';
+
+class App extends React.Component<{}, {}> {
+  render(): ReactNode {
+    return (
+        <div>
+          <Board />
+        </div>
+    );
+  }
+}
+
 
 ReactDOM.render(<App />, document.getElementById('root'));
